perf_dashboard/requirements.txt
Django==4.1.3
beautifulsoup4==4.8.1
pandas==1.5.0
wget==3.2
whitenoise==5.0.1
google-cloud-storage==2.6.0
tzdata==2022.7
